chore: do some suggestions
