+++ packages/core/src/node/runtimeModule/siteData/extractPageData.ts
@@ -1,18 +1,18 @@
 import path from 'path';
-import {
-  PageIndexInfo,
-  ReplaceRule,
-  Header,
-  MDX_REGEXP,
-} from '@rspress/shared';
-import { htmlToText } from 'html-to-text';
 import fs from '@rspress/shared/fs-extra';
+import { htmlToText } from 'html-to-text';
 import { compile } from '@rspress/mdx-rs';
 import { loadFrontMatter } from '@rspress/shared/node-utils';
+import {
+  Header,
+  MDX_REGEXP,
+  ReplaceRule,
+  PageIndexInfo,
+} from '@rspress/shared';
 import { flattenMdxContent } from '@/node/utils';
+import { importStatementRegex } from '@/node/constants';
 import { RouteService } from '@/node/route/RouteService';
 import { applyReplaceRules } from '@/node/utils/applyReplaceRules';
-import { importStatementRegex } from '@/node/constants';
 
 export async function extractPageData(
   replaceRules: ReplaceRule[],
@@ -78,7 +78,6 @@ export async function extractPageData(
         return null;
       }
 
-      languages.forEach(lang => highlighterLangs.add(lang));
       content = htmlToText(String(html), {
         wordwrap: 80,
         selectors: [
@@ -114,6 +113,7 @@ export async function extractPageData(
         content = content.slice(title.length);
       }
 
+      languages.forEach(lang => highlighterLangs.add(lang));
       const toc: Header[] = rawToc.map(item => {
         // If the item.id ends with '-number', we take the number
         const match = item.id.match(/-(\d+)$/);
